+++ src/constants/projectsData/index.js
@@ -1,3 +1,73 @@
+import planeImg from "assets/images/nProjects/plane.png"
+import restaurantReactImg from "assets/images/nProjects/restaurant_react.png"
+import watchesImg from "assets/images/nProjects/watches.png"
+import blogImg from "assets/images/nProjects/blog.png"
+import realtorImg from "assets/images/nProjects/realtor.png"
+import beats2Img from "assets/images/nProjects/beats2.png"
+
+export const projectsData = [
+    {
+        title: "Real Estate",
+        description:
+            "Find your dream home with our extensive collection of real estate listings.",
+        img: realtorImg,
+        imgDesktop: realtorImg,
+        technologies: ["HTML", "JavaScript", "CSS"],
+        visit: "https://real-estate-sable-six.vercel.app/",
+        id: 8,
+    },
+    {
+        title: "Headphones",
+        description:
+            "Immerse yourself in high-quality sound with our premium noise-cancelling headphones, designed for comfort and performance.",
+        img: beats2Img,
+        imgDesktop: beats2Img,
+        technologies: ["HTML", "JavaScript", "CSS"],
+        visit: "https://ecommerce-art1us.vercel.app/",
+        id: 3,
+    },
+    {
+        title: "Flight Booking App",
+        description:
+            "Book your next adventure with our easy-to-use flight booking app, designed to help you find the best deals and plan your travel itinerary with ease.",
+        img: planeImg,
+        imgDesktop: planeImg,
+        technologies: ["HTML", "JavaScript", "CSS"],
+        visit: "https://unrivaled-douhua-a237e4.netlify.app/",
+        id: 12,
+    },
+    {
+        title: "Restaurant",
+        description:
+            "Experience an unforgettable dining journey at our restaurant, where we use the freshest and highest quality ingredients to craft a menu that is both diverse and delectable.",
+        img: restaurantReactImg,
+        imgDesktop: restaurantReactImg,
+        technologies: ["React"],
+        visit: "https://stirring-phoenix-05f76a.netlify.app/",
+        id: 13,
+    },
+    {
+        title: "Watches",
+        description:
+            "Elevate your fashion game with our exquisite collection of watches, crafted with precision and designed to make a bold statement – shop now to find your next favorite accessory.",
+        img: watchesImg,
+        imgDesktop: watchesImg,
+        technologies: ["HTML", "JavaScript", "CSS"],
+        visit: "https://a-s-t-r-a.github.io/responsive-watches-website/",
+        id: 14,
+    },
+    {
+        title: "Blog website",
+        description:
+            "Stay informed and inspired with our engaging and informative blog, featuring a variety of topics and perspectives that cater to your interests – explore now to expand your knowledge and curiosity.",
+        img: blogImg,
+        imgDesktop: blogImg,
+        technologies: ["HTML", "JavaScript", "CSS"],
+        visit: "https://astra-blog.vercel.app/",
+        id: 18,
+    },
+]
+
 /* import {
     projImg1_1,
     projImg1_2,
