--- src/modules/NProjects/data.js
@@ -1,69 +1,0 @@
-import plane from "assets/images/nProjects/plane.png"
-import restaurant_react from "assets/images/nProjects/restaurant_react.png"
-import watches from "assets/images/nProjects/watches.png"
-import blog from "assets/images/nProjects/blog.png"
-import realtor from "assets/images/nProjects/realtor.png"
-import beats2 from "assets/images/nProjects/beats2.png"
-
-export const projectsData = [
-    {
-        title: "Real Estate",
-        description:
-            "Find your dream home with our extensive collection of real estate listings.",
-        img: realtor,
-        imgDesktop: realtor,
-        technologies: ["HTML", "JavaScript", "CSS"],
-        visit: "https://real-estate-sable-six.vercel.app/",
-        id: 8,
-    },
-    {
-        title: "Headphones",
-        description:
-            "Immerse yourself in high-quality sound with our premium noise-cancelling headphones, designed for comfort and performance.",
-        img: beats2,
-        imgDesktop: beats2,
-        technologies: ["HTML", "JavaScript", "CSS"],
-        visit: "https://ecommerce-art1us.vercel.app/",
-        id: 3,
-    },
-    {
-        title: "Flight Booking App",
-        description:
-            "Book your next adventure with our easy-to-use flight booking app, designed to help you find the best deals and plan your travel itinerary with ease.",
-        img: plane,
-        imgDesktop: plane,
-        technologies: ["HTML", "JavaScript", "CSS"],
-        visit: "https://unrivaled-douhua-a237e4.netlify.app/",
-        id: 12,
-    },
-    {
-        title: "Restaurant",
-        description:
-            "Experience an unforgettable dining journey at our restaurant, where we use the freshest and highest quality ingredients to craft a menu that is both diverse and delectable.",
-        img: restaurant_react,
-        imgDesktop: restaurant_react,
-        technologies: ["React"],
-        visit: "https://stirring-phoenix-05f76a.netlify.app/",
-        id: 13,
-    },
-    {
-        title: "Watches",
-        description:
-            "Elevate your fashion game with our exquisite collection of watches, crafted with precision and designed to make a bold statement – shop now to find your next favorite accessory.",
-        img: watches,
-        imgDesktop: watches,
-        technologies: ["HTML", "JavaScript", "CSS"],
-        visit: "https://a-s-t-r-a.github.io/responsive-watches-website/",
-        id: 14,
-    },
-    {
-        title: "Blog website",
-        description:
-            "Stay informed and inspired with our engaging and informative blog, featuring a variety of topics and perspectives that cater to your interests – explore now to expand your knowledge and curiosity.",
-        img: blog,
-        imgDesktop: blog,
-        technologies: ["HTML", "JavaScript", "CSS"],
-        visit: "https://astra-blog.vercel.app/",
-        id: 18,
-    },
-]
