+++ src/modules/NProjects/ui/components/ProjectImage/ProjectImage.module.css
@@ -11,5 +11,5 @@
 }
 
 .img.active {
-    filter: brightness(100%);
+    filter: brightness(90%);
 }
